CadreAIAssesment/requirements.txt
openai>=1.0.0
PyMuPDF>=1.23.7
python-dotenv
